+++ run.sh
@@ -6,5 +6,5 @@ if [ -f /etc/varnish/vcl.d/*.vcl ] ; then
 fi
 
 # Start varnish and log
-varnishd -f /etc/varnish/default.vcl -s file,/var/lib/varnish/varnish_storage.bin,${CACHE_SIZE} -T 0.0.0.0:${VARNISH_ADMIN_PORT} -a 0.0.0.0:${VARNISH_PORT} ${VARNISHD_PARAMS}
-varnishlog
+varnishd -f /etc/varnish/default.vcl -s malloc,${CACHE_SIZE} -S /etc/varnish/secret -T 0.0.0.0:${VARNISH_ADMIN_PORT} -a 0.0.0.0:${VARNISH_PORT} ${VARNISHD_PARAMS} -n ${INSTANCE}
+varnishlog -n ${INSTANCE}