+++ CHANGELOG.md
@@ -6,6 +6,9 @@ This file is used to list changes made in each version of docker-varnish.
 
 * add import std
 * allow additional vcl configs in /etc/varnish/vcl.d
+* extend docker-compose example
+* add definition of varnish secret
+* use malloc instead of file backend by default
 
 ## 1.0.1 (December 5, 2015)